Modifying due to unknown bug

+++ src/Catalog.Api/Products/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,28 +1,28 @@
-    using Catalog.Api.Exceptions;
-    using Catalog.Api.Models;
+using Catalog.Api.Exceptions;
+using Catalog.Api.Models;
 
-    namespace Catalog.Api.Products.DeleteProduct;
+namespace Catalog.Api.Products.DeleteProduct;
 
-    public sealed record DeleteProductCommand(Guid Id)
-        : ICommand<DeleteProductResult>;
+public sealed record DeleteProductCommand(Guid Id)
+    : ICommand<DeleteProductResult>;
 
-    public record DeleteProductResult(bool isSuccess);
+public record DeleteProductResult(bool isSuccess);
 
-    public class DeleteProductCommandHandler(IDocumentSession session, ILogger<DeleteProductCommandHandler> logger)
-        : ICommandHandler<DeleteProductCommand, DeleteProductResult>
+public class DeleteProductCommandHandler(IDocumentSession session, ILogger<DeleteProductCommandHandler> logger)
+    : ICommandHandler<DeleteProductCommand, DeleteProductResult>
+{
+    public async Task<DeleteProductResult> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
     {
-        public async Task<DeleteProductResult> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
-        {
-            logger.LogInformation("DeleteProductCommandHandler.Handle was called {command}", command);
+        logger.LogInformation("DeleteProductCommandHandler.Handle was called {command}", command);
 
-            // var product = await session.LoadAsync<Product>(command.Id, cancellationToken);
-            // if (product is null)
-            // {
-            //     throw new ProductNotFoundException();
-            // }
-            session.Delete<Product>(command.Id);
-            await session.SaveChangesAsync(cancellationToken);
-            
-            return new DeleteProductResult(true);
+        var product = await session.LoadAsync<Product>(command.Id, cancellationToken);
+        if (product is null)
+        {
+            throw new ProductNotFoundException();
         }
+        session.Delete<Product>(command.Id);
+        await session.SaveChangesAsync(cancellationToken);
+        
+        return new DeleteProductResult(true);;
     }
+}
